requirements.txt
Django==3.2
djangorestframework==3.13.1
mysqlclient==2.2.7
djangorestframework-simplejwt
django-cors-headers
gunicorn
